requirements.txt
streamlit==1.43.2
openai==1.68.2
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.4.2
networkx==3.3
pypdf==5.3.1
python-docx==1.1.2
tiktoken==0.9.0
faiss-cpu==1.8.0.post1
rdflib==7.1.3
pyvis==0.3.2
matplotlib==3.8.4